Gemini Project/requirements.txt
streamlit
requests
beautifulsoup4
faiss-cpu
langchain
langchain-community
google-generativeai
python-dotenv


